requirements.txt
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.1.0
autobahn==23.6.2
Automat==22.10.0
certifi==2023.11.17
cffi==1.16.0
channels==4.0.0
channels-redis==4.1.0
colorama==0.4.6
constantly==23.10.4
coverage==7.3.4
cryptography==41.0.5
daphne==4.0.0
Django==4.2.5
django-pytest==0.2.0
django-redis==5.4.0
exceptiongroup==1.2.0
factory-boy==3.3.0
Faker==21.0.0
h11==0.14.0
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
inflection==0.5.1
iniconfig==2.0.0
msgpack==1.0.7
mysql==0.0.3
mysql-connector-python==8.1.0
mysqlclient==2.2.0
outcome==1.3.0.post0
packaging==23.2
Pillow==10.0.1
pluggy==1.3.0
protobuf==4.21.12
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pyOpenSSL==23.3.0
PySocks==1.7.1
pytest==7.4.3
pytest-cov==4.1.0
pytest-django==4.7.0
pytest-factoryboy==2.6.0
python-dateutil==2.8.2
python-decouple==3.8
redis==5.0.1
selenium==4.16.0
service-identity==23.1.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
sqlparse==0.4.4
tomli==2.0.1
trio==0.23.2
trio-websocket==0.11.1
Twisted==23.10.0
twisted-iocpsupport==1.0.4
txaio==23.1.1
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.1.0
wsproto==1.2.0
zope.interface==6.1
